requirement.txt
streamlit 
tensorflow
pillow 
openai 
langchain
